pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.3.2"
description = "Difference Calculator"
authors = ["Ivan <kozel.vania@yandex.ru>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest-cov = "^3.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pycodestyle = "^2.8.0"
wemake-python-styleguide = "^0.16.1"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
